# Track evictions in the PostingsForMatchers cache