In https://github.com/grafana/mimir-prometheus/pull/822 I've introduced metrics for the `PostingsForMatchers` cache. In this PR I'm extending the metrics to count evictions by reason.

Ref: https://github.com/grafana/pir-actions/issues/307

**Benchmarks**

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Apple M3 Pro
                                                          │ before.txt  │             after.txt             │
                                                          │   sec/op    │   sec/op     vs base              │
PostingsForMatchersCache/no_evictions-11                    639.1n ± 2%   636.2n ± 1%       ~ (p=0.394 n=6)
PostingsForMatchersCache/high_eviction_rate-11              10.66µ ± 0%   10.70µ ± 1%       ~ (p=0.071 n=6)
PostingsForMatchersCache_ConcurrencyOnHighEvictionRate-11   328.2n ± 1%   326.3n ± 3%       ~ (p=0.167 n=6)
geomean                                                     1.308µ        1.305µ       -0.24%

                                                          │  before.txt  │              after.txt               │
                                                          │     B/op     │     B/op      vs base                │
PostingsForMatchersCache/no_evictions-11                      974.0 ± 0%     974.0 ± 0%       ~ (p=1.000 n=6) ¹
PostingsForMatchersCache/high_eviction_rate-11              26.32Ki ± 0%   26.32Ki ± 0%       ~ (p=1.000 n=6) ¹
PostingsForMatchersCache_ConcurrencyOnHighEvictionRate-11   1.000Ki ± 0%   1.000Ki ± 0%       ~ (p=1.000 n=6)
geomean                                                     2.925Ki        2.925Ki       +0.00%
¹ all samples are equal

                                                          │ before.txt │             after.txt              │
                                                          │ allocs/op  │ allocs/op   vs base                │
PostingsForMatchersCache/no_evictions-11                    20.00 ± 0%   20.00 ± 0%       ~ (p=1.000 n=6) ¹
PostingsForMatchersCache/high_eviction_rate-11              46.00 ± 0%   46.00 ± 0%       ~ (p=1.000 n=6) ¹
PostingsForMatchersCache_ConcurrencyOnHighEvictionRate-11   20.00 ± 0%   20.00 ± 5%       ~ (p=1.000 n=6)
geomean                                                     26.40        26.40       +0.00%
¹ all samples are equal
```